Implement stage
The stage is implemented via a static tileMap in tileMap.lua,
which exports only a matrix of tile IDs.

Some bugs with the player model were fixed.

Tile model is currently an interface; it might be useful for
collision detection wtih the player model.

Config.lua is created for configuring LOVE2D's default stuff
before the game is loaded.